[ refactor ] : Kafka 연결  Refactoring
