+++ server-java/src/test/java/kr/hhplus/be/server/kafka/ConnectionTest.java
@@ -1,18 +1,18 @@
 package kr.hhplus.be.server.kafka;
 
-import kr.hhplus.be.server.common.config.kafka.KafkaPlatformConsumer;
-import kr.hhplus.be.server.common.config.kafka.ReservationProducer;
+import kr.hhplus.be.server.interfaces.controller.reservation.KafkaPlatformConsumer;
+import kr.hhplus.be.server.infrastructure.reservation.ReservationProducer;
+import kr.hhplus.be.server.domain.user.User;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.kafka.annotation.EnableKafka;
-
-import java.util.HashMap;
+import org.testcontainers.shaded.com.fasterxml.jackson.core.JsonProcessingException;
+import org.testcontainers.shaded.com.fasterxml.jackson.databind.ObjectMapper;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @SpringBootTest
 @EnableKafka
@@ -27,18 +27,16 @@ public class ConnectionTest {
 
     @DisplayName("Kafka 에서 메세지 보내면, 해당 메세지가 전송이 된다.")
     @Test
-    void sendToKafka_00() throws InterruptedException {
+    void sendToKafka_00() throws InterruptedException, JsonProcessingException {
         // given
-        HashMap<String, String> exerciseMap = new HashMap<>();
-        exerciseMap.put("name", "kim");
-        exerciseMap.put("age", "16");
+        User user = User.create("TEST STRING");
 
         // when
-        reservationProducer.sendReservation("my-first-topic", exerciseMap);
+        reservationProducer.sendReservation("my-first-topic", user);
 
-        boolean isMessageReceived = kafkaConsumer.awaitLatch();
-
+        kafkaConsumer.awaitLatch();
         String receivedMessage = kafkaConsumer.getReceivedMessage();
-        assertEquals("{name=kim, age=16}", receivedMessage);
+
+        assertEquals(new ObjectMapper().writeValueAsString(user),receivedMessage);
     }
 }
